refactor: class QueryFormatBuilder extends SelectQueryBuilder. QueryFormatter has limited properties of QueryFormatBuilder
